Template Files
Added standard templates for some of the file types (models, views,
controllers, channels, tests).
These are used to create a file when you do `:Emodel foo` and foo.ex
does not exist, or if you switch to an alternate that does not exist
using `:A`.

Changed the `autocmds` that phoenix.vim sets up on.